Move test to ReactContextValidator
